+++ src/renderer.ts
@@ -1,15 +1,33 @@
-import { app, BrowserWindow } from 'electron';
+import { app, BrowserWindow, shell } from 'electron';
 
 function createWindow() {
   const mainWindow = new BrowserWindow({
     width: 800,
-    height: 600
+    height: 600,
+    show: false
   });
 
   const params = process.argv.slice(2);
   mainWindow.loadURL('http://localhost:' + params[0]);
+
+  mainWindow.once('ready-to-show', () => mainWindow.show());
 }
 
 app.whenReady().then(() => {
   createWindow();
+});
+
+app.on('web-contents-created', (_, contents) => {
+  // Angular router is ignored on `will-navigate` event
+  contents.on('will-navigate', event => {
+    event.preventDefault();
+  });
+
+  contents.setWindowOpenHandler(({ url }) => {
+    // open all blank href links using the OS default browser
+    setImmediate(() => {
+      shell.openExternal(url);
+    });
+    return { action: 'deny' };
+  });
 });